Phone add process created

+++ patikaStore/src/com/company/PhoneProcess.java
@@ -3,7 +3,7 @@ package com.company;
 public class PhoneProcess extends Product{
     private double batteryPower;
     private String color;
-    public PhoneProcess(int id, String productName, String categoryName, int stockAmount, String brand, int price, double discountRate, int ram, int memory, int batteryPower, String color, double screenSize) {
+    public PhoneProcess(int id, String productName, String categoryName, int stockAmount, String brand, int price, double discountRate, int ram, int memory, double batteryPower, String color, double screenSize) {
         super(id, productName, categoryName, stockAmount, brand, price, discountRate, ram, memory, screenSize);
         this.batteryPower = batteryPower;
         this.color = color;
